move and refine tests
