+++ core/tests/behavior/main.rs
@@ -28,7 +28,6 @@ mod async_delete;
 mod async_list;
 mod async_presign;
 mod async_read;
-mod async_remove_all;
 mod async_rename;
 mod async_stat;
 mod async_write;
@@ -39,7 +38,6 @@ mod blocking_create_dir;
 mod blocking_delete;
 mod blocking_list;
 mod blocking_read;
-mod blocking_remove_all;
 mod blocking_rename;
 mod blocking_stat;
 mod blocking_write;
@@ -73,7 +71,6 @@ fn main() -> anyhow::Result<()> {
     async_rename::tests(&op, &mut tests);
     async_stat::tests(&op, &mut tests);
     async_write::tests(&op, &mut tests);
-    async_remove_all::tests(&op, &mut tests);
 
     blocking_copy::tests(&op, &mut tests);
     blocking_create_dir::tests(&op, &mut tests);
@@ -83,7 +80,6 @@ fn main() -> anyhow::Result<()> {
     blocking_rename::tests(&op, &mut tests);
     blocking_stat::tests(&op, &mut tests);
     blocking_write::tests(&op, &mut tests);
-    blocking_remove_all::tests(&op, &mut tests);
 
     // Don't init logging while building operator which may break cargo
     // nextest output
